requirements.txt
celery==5.4.0
beautifulsoup4==4.12.3
email-validator==2.2.0
flake8==7.1.1
flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
flask-login==0.6.3
flask-migrate==4.0.7
redis==5.0.8
requests==2.32.3
